kubeflowTalk/pyproject.toml
[tool.poetry]
name = "vertexai-pipelines"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"
packages = [{include = "vertexai_pipelines"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12.0"
google-cloud-aiplatform = "^1.33.1"
kfp = "2.0.0"
google-cloud-pipeline-components = "^2.3.1"
google-cloud = "^0.34.0"
numpy = "^1.26.0"
pillow = "^10.0.1"
tqdm = "^4.66.1"
pandas = "^2.1.1"
scikit-learn = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
